ai-llm-sraping/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
beautifulsoup4
lxml
openai==0.28
faiss-cpu
google-cloud-storage
pytest
httpx
isort
